tarabbit_server/requirements.txt
anyio==3.7.1
fastapi==0.75.0
openai==1.30.4
pymongo==4.7.2
python-dotenv==0.19.2
pytz==2024.1
requests==2.31.0
setuptools==69.5.1
urllib3==1.26.15
uvicorn==0.17.0
APScheduler==3.10.4
cryptography==42.0.8
pyopenssl==24.1.0